Fixed bug with progress dialog displaying incorrect max
Created new AsyncTask for getting the AuthToken
Created new AsyncTask for getting mail before parsing
Fixed a bug with not properly returning the AuthToken if it was cached.
EmailParseTask is now created with all the Ingress emails already
loaded, so we can initialize the progress dialog with the correct
maximum before displaying.